// src/app/models/course.ts
export class Course {
    id: number | any;
    name: string | undefined;
    summary: string | undefined;
    description: string | undefined;
    sequence: string | undefined;
    imageUrl: string | undefined;
    demoUrl: string | undefined;
    url: string | undefined;
    unitPrice: number | undefined;
    difficultyType: number | undefined;
    createdDate: Date | undefined;
    updatedDate: Date | undefined;
    isActive: boolean | undefined;
    categoryId: number | undefined;
    mentorId: number | undefined;
}